move Faq.text.js to config folder, change name
