+++ src/config/faqs.js
